(temporary) Fix compiler error in updater

+++ UI/win-update/updater/hash.cpp
@@ -59,7 +59,7 @@ bool CalculateFileHash(const wchar_t *path, BYTE *hash)
 
 	for (;;) {
 		DWORD read = 0;
-		if (!ReadFile(handle, &hashBuffer[0], hashBuffer.size(), &read,
+		if (!ReadFile(handle, &hashBuffer[0], (DWORD)hashBuffer.size(), &read,
 			      nullptr))
 			return false;
